backend: split layout code into own subpackage
