+++ internal/backend/layout/layout.go
@@ -1,4 +1,4 @@
-package backend
+package layout
 
 import (
 	"context"
