+++ internal/backend/layout/layout_default.go
@@ -1,4 +1,4 @@
-package backend
+package layout
 
 import (
 	"encoding/hex"
